Removed required indicator
Reason for removal is based on a likely scenario where publication exclusively has non-DPSM authors, meaning this field would be null (i.e. leaving it required would prevent a form submission). A reverse scenario where publication exclusively has DPSM authors accounts for why the non-DPSM authors field is also not required.
Best course of action for form validation is some sort of method that checks if both fields are null (which is the only invalid scenario).